+++ README.md
@@ -1,0 +1,14 @@
+## run 所有的case
+根目录下执行
+```
+pytest 
+```
+
+## 执行单个文件
+根目录下执行
+```
+pytest testcase\test_msg.py
+```
+
+### conftest.py
+将包含共享的方法放到 conftest.py 中 运行过程中会自动加载
